完成了文章收藏的改进和测试

+++ src/main/java/yellow/iblog/model/ArticleLikeResponse.java
@@ -7,6 +7,6 @@ public class ArticleLikeResponse {
     //用户当前是否点赞了文章
     private Boolean status;
     //文章当前点赞数
-    private long crtLikes;
+    private int crtLikes;
 
 }
